Added support to send customized JSON for web-server
Altered the format of the JSON to match that of the database schema we are using. We are formatting it as an array called read info, and putting all the sensors in there wih an id, type and value. I am also trying to get the device to read the response back from the server too.